Conflicts:
	visualisation/css/layout.css

+++ visualisation/requests/table_clear.php
@@ -1,11 +1,13 @@
 <?php
+	session_start();
+	include('../data/mysql.php');
 
-include('../data/mysql.php');
+	if($_SESSION['userid']>0){
 
-$table = $_POST['table'];
-echo $table;
+		$table = $_POST['table'];
+		echo $table;
 
-$result = mysql_query("TRUNCATE $table") or die('Error: ' . mysql_error());
+		$result = mysql_query("TRUNCATE $table") or die('Error: ' . mysql_error());
 
-
+	}
 ?>
